Split exec process into local and remote implementations
Stack this on top of PR #15232's filesystem split: expose the process trait on Environment, keep LocalProcess as the real implementation, proxy RemoteProcess over ExecServerClient, and route RPCs through a thin ProcessHandler.
