+++ codex-rs/exec-server/src/client_api.rs
@@ -1,8 +1,5 @@
 use std::time::Duration;
 
-use crate::protocol::ExecExitedNotification;
-use crate::protocol::ExecOutputDeltaNotification;
-
 /// Connection options for any exec-server client transport.
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct ExecServerClientConnectOptions {
@@ -17,11 +14,4 @@ pub struct RemoteExecServerConnectArgs {
     pub client_name: String,
     pub connect_timeout: Duration,
     pub initialize_timeout: Duration,
-}
-
-/// Connection-level server events.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum ExecServerEvent {
-    OutputDelta(ExecOutputDeltaNotification),
-    Exited(ExecExitedNotification),
 }
