+++ codex-rs/exec-server/src/environment.rs
@@ -1,15 +1,34 @@
+use std::sync::Arc;
+
 use crate::ExecServerClient;
 use crate::ExecServerError;
 use crate::RemoteExecServerConnectArgs;
 use crate::file_system::ExecutorFileSystem;
 use crate::local_file_system::LocalFileSystem;
+use crate::local_process::LocalProcess;
+use crate::process::ExecProcess;
 use crate::remote_file_system::RemoteFileSystem;
-use std::sync::Arc;
+use crate::remote_process::RemoteProcess;
 
-#[derive(Clone, Default)]
+pub trait ExecutorEnvironment: Send + Sync {
+    fn get_executor(&self) -> Arc<dyn ExecProcess>;
+}
+
+#[derive(Clone)]
 pub struct Environment {
     experimental_exec_server_url: Option<String>,
     remote_exec_server_client: Option<ExecServerClient>,
+    executor: Arc<dyn ExecProcess>,
+}
+
+impl Default for Environment {
+    fn default() -> Self {
+        Self {
+            experimental_exec_server_url: None,
+            remote_exec_server_client: None,
+            executor: Arc::new(LocalProcess::default()),
+        }
+    }
 }
 
 impl std::fmt::Debug for Environment {
@@ -19,11 +38,7 @@ impl std::fmt::Debug for Environment {
                 "experimental_exec_server_url",
                 &self.experimental_exec_server_url,
             )
-            .field(
-                "has_remote_exec_server_client",
-                &self.remote_exec_server_client.is_some(),
-            )
-            .finish()
+            .finish_non_exhaustive()
     }
 }
 
@@ -31,22 +46,38 @@ impl Environment {
     pub async fn create(
         experimental_exec_server_url: Option<String>,
     ) -> Result<Self, ExecServerError> {
-        let remote_exec_server_client =
-            if let Some(websocket_url) = experimental_exec_server_url.as_deref() {
-                Some(
-                    ExecServerClient::connect_websocket(RemoteExecServerConnectArgs::new(
-                        websocket_url.to_string(),
-                        "codex-core".to_string(),
-                    ))
-                    .await?,
-                )
-            } else {
-                None
-            };
+        let remote_exec_server_client = if let Some(url) = &experimental_exec_server_url {
+            Some(
+                ExecServerClient::connect_websocket(RemoteExecServerConnectArgs {
+                    websocket_url: url.clone(),
+                    client_name: "codex-environment".to_string(),
+                    connect_timeout: std::time::Duration::from_secs(5),
+                    initialize_timeout: std::time::Duration::from_secs(5),
+                })
+                .await?,
+            )
+        } else {
+            None
+        };
+
+        let executor: Arc<dyn ExecProcess> = if let Some(client) = remote_exec_server_client.clone()
+        {
+            Arc::new(RemoteProcess::new(client))
+        } else {
+            let local_process = LocalProcess::default();
+            local_process
+                .initialize()
+                .map_err(|err| ExecServerError::Protocol(err.message))?;
+            local_process
+                .initialized()
+                .map_err(ExecServerError::Protocol)?;
+            Arc::new(local_process)
+        };
 
         Ok(Self {
             experimental_exec_server_url,
             remote_exec_server_client,
+            executor,
         })
     }
 
@@ -54,8 +85,12 @@ impl Environment {
         self.experimental_exec_server_url.as_deref()
     }
 
-    pub fn remote_exec_server_client(&self) -> Option<&ExecServerClient> {
-        self.remote_exec_server_client.as_ref()
+    pub fn remote_exec_server_client(&self) -> Option<ExecServerClient> {
+        self.remote_exec_server_client.clone()
+    }
+
+    pub fn get_executor(&self) -> Arc<dyn ExecProcess> {
+        Arc::clone(&self.executor)
     }
 
     pub fn get_filesystem(&self) -> Arc<dyn ExecutorFileSystem> {
@@ -64,6 +99,12 @@ impl Environment {
         } else {
             Arc::new(LocalFileSystem)
         }
+    }
+}
+
+impl ExecutorEnvironment for Environment {
+    fn get_executor(&self) -> Arc<dyn ExecProcess> {
+        self.get_executor()
     }
 }
