+++ codex-rs/exec-server/src/lib.rs
@@ -4,15 +4,17 @@ mod connection;
 mod environment;
 mod file_system;
 mod local_file_system;
+mod local_process;
+mod process;
 mod protocol;
 mod remote_file_system;
+mod remote_process;
 mod rpc;
 mod server;
 
 pub use client::ExecServerClient;
 pub use client::ExecServerError;
 pub use client_api::ExecServerClientConnectOptions;
-pub use client_api::ExecServerEvent;
 pub use client_api::RemoteExecServerConnectArgs;
 pub use codex_app_server_protocol::FsCopyParams;
 pub use codex_app_server_protocol::FsCopyResponse;
@@ -20,7 +22,6 @@ pub use codex_app_server_protocol::FsCreateDirectoryParams;
 pub use codex_app_server_protocol::FsCreateDirectoryResponse;
 pub use codex_app_server_protocol::FsGetMetadataParams;
 pub use codex_app_server_protocol::FsGetMetadataResponse;
-pub use codex_app_server_protocol::FsReadDirectoryEntry;
 pub use codex_app_server_protocol::FsReadDirectoryParams;
 pub use codex_app_server_protocol::FsReadDirectoryResponse;
 pub use codex_app_server_protocol::FsReadFileParams;
@@ -30,6 +31,7 @@ pub use codex_app_server_protocol::FsRemoveResponse;
 pub use codex_app_server_protocol::FsWriteFileParams;
 pub use codex_app_server_protocol::FsWriteFileResponse;
 pub use environment::Environment;
+pub use environment::ExecutorEnvironment;
 pub use file_system::CopyOptions;
 pub use file_system::CreateDirectoryOptions;
 pub use file_system::ExecutorFileSystem;
@@ -37,6 +39,8 @@ pub use file_system::FileMetadata;
 pub use file_system::FileSystemResult;
 pub use file_system::ReadDirectoryEntry;
 pub use file_system::RemoveOptions;
+pub use process::ExecProcess;
+pub use process::ExecServerEvent;
 pub use protocol::ExecExitedNotification;
 pub use protocol::ExecOutputDeltaNotification;
 pub use protocol::ExecOutputStream;
