+++ codex-rs/exec-server/src/server.rs
@@ -1,5 +1,6 @@
 mod file_system_handler;
 mod handler;
+mod process_handler;
 mod processor;
 mod registry;
 mod transport;
